% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scluster.R
\name{sclDataConstructor}
\alias{sclDataConstructor}
\title{sclData Constructor}
\usage{
sclDataConstructor(tags, tagType = "raw")
}
\arguments{
\item{tags}{A matrix of tags where the rows correspond to features (genes, transcripts, etc) and the columns correspond to cells.}

\item{tagType}{- \code{raw} for when tags are raw tags ; \code{cpm} when tags are tags per million ; default is \code{raw}.}
}
\value{
an sclData class object.
}
\description{
\code{sclDataConstructor} creates a new sclData class object from a tag table.
}
\details{
Creates an object in sclData (single-cell RNA-Seq dataset) class.
Attributes of the class include scalar, vector and matrix
data types necessary for the \emph{sclust} analysis - such as tag table, 
library  sizes, dropout candidates, imputation weighting threshold. The 
tags can be  raw tags (default) or tags per million (cpm).  Raw tags 
are preferrable as the individual library sizes, as determined by the raw 
tags, are used to determine dropout candidates.
}
\examples{
example(scluster)
}

