% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scluster.R
\docType{methods}
\name{sclust}
\alias{sclust}
\title{sclust}
\usage{
\S4method{sclust}{sclData}(object, nCluster, s = NA_real_, nf = NA_integer_,
  max.nIter = 10, markers, supervise = F, alpha, seed = 123,
  dist = "euclidean", priorTPM = 1)
}
\arguments{
\item{nCluster}{Number of clusters.}

\item{s}{A tuning parameter determines the sum of weights for all features.  NOTE that only one of the parameters \code{nf} and \code{s} should be set; if neither are set, then a default value of \code{nf=300} is applied.}

\item{nf}{Determines the number of features (with non-zero weight) used for clustering.}

\item{max.nIter}{The number of iterations.}

\item{markers}{A list of names of marker genes, not used when supervise=FALSE.}

\item{supervise}{Boolean; If TRUE, implement supervised clustering.}

\item{alpha}{A tuning parameter determines the sum of weights of marker genes.}

\item{seed}{Used for random number generation reproducability.}

\item{dist}{Either "ecuclidean" or "cidr".  If "euclidean", then the squared euclidean distance is used, otherwise the CIDR dissimilarity measure is used.}

\item{priorTPM}{Count to add to deal with 0 tags for log function.}
}
\value{
An updated sclData object with the following additional attributes:

\item{cluster}{A vector with the clustering result.}
\item{weight}{nIter x p matrix; the weight for each feature in each iteration.}
\item{w.bcss}{The weighted between cluster sum of squares in each iteration.}
\item{bcss}{nIter x p matrix; the between cluster sum of squares for each feature in each iteration.}
\item{totFeatDist}{1 x p matrix; the total distance calculated for each feature.}
\item{sumWeight}{nIter x 1 matrix; the sum of weight in each iteration.}
\item{nf.zw}{nIter x 1 matrix; the number of features with non-zero weight in each iteration.}
}
\examples{
example(scluster)
}

